+++ packages/react-menu/src/index.ts
@@ -2,3 +2,5 @@ import './version';
 
 export * from './MenuItem';
 export * from './MenuList';
+export * from './MenuItemCheckbox';
+export * from './MenuItemRadio';